Adds Guard for koan-esque progress through specs
Guard is configured to run RSpec in "fail-fast" mode, which means it
ends as soon as it hits a failing spec. Now users don't have to
uncomment each spec as they go through the exercise, but at the same
time won't be overwhelmed by a ton of red specs.

I also restructured the directory so you don't need to keep making spec_helper
files for each exercise/theme